move reap list metrics into its internal package
